requirements.txt
transformers==4.37.0
datasets
accelerate==0.27.2
deepspeed==0.13.2
evaluate
matplotlib
hydra-core
omegaconf
peft
rouge_score
tqdm
matplotlib
einops
bitsandbytes
scipy
ninja
openai
spacy
nltk
natsort
stanza==1.8.1